requirements.txt
fastapi
uvicorn
openai
feedparser
requests
python-dotenv
langchain-community
fuzzywuzzy